Added TTL for job logs in elasticsearch. Job logs are deleted from Elasticsearch when job is deleted. Refactored ElasticService.
