+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/elastic/ElasticHit.java
@@ -1,4 +1,4 @@
-package io.hops.hopsworks.api.elastic;
+package io.hops.hopsworks.common.elastic;
 
 import java.util.HashMap;
 import java.util.Map;
