+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/jobs/execution/ExecutionController.java
@@ -5,19 +5,24 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import io.hops.hopsworks.common.dao.user.activity.ActivityFacade;
 import io.hops.hopsworks.common.dao.hdfs.inode.Inode;
 import io.hops.hopsworks.common.dao.hdfs.inode.InodeFacade;
 import io.hops.hopsworks.common.dao.jobhistory.Execution;
+import io.hops.hopsworks.common.dao.jobhistory.YarnApplicationstateFacade;
 import io.hops.hopsworks.common.dao.jobs.JobsHistoryFacade;
 import io.hops.hopsworks.common.dao.jobs.description.JobDescription;
 import io.hops.hopsworks.common.dao.user.Users;
+import io.hops.hopsworks.common.hdfs.DistributedFileSystemOps;
+import io.hops.hopsworks.common.hdfs.DistributedFsService;
+import io.hops.hopsworks.common.hdfs.HdfsUsersController;
 import io.hops.hopsworks.common.jobs.adam.AdamController;
 import io.hops.hopsworks.common.jobs.flink.FlinkController;
 import io.hops.hopsworks.common.jobs.spark.SparkController;
 import io.hops.hopsworks.common.jobs.spark.SparkJobConfiguration;
+import io.hops.hopsworks.common.util.Settings;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  * Takes care of booting the execution of a job.
@@ -38,8 +43,16 @@ public class ExecutionController {
   private ActivityFacade activityFacade;
   @EJB
   private JobsHistoryFacade jobHistoryFac;
+  @EJB
+  private YarnApplicationstateFacade yarnApplicationstateFacade;
+  @EJB
+  private HdfsUsersController hdfsUsersBean;
+  @EJB
+  private DistributedFsService dfs;
+  @EJB
+  private Settings settings;
 
-  final Logger logger = LoggerFactory.getLogger(ExecutionController.class);
+  final Logger logger = Logger.getLogger(ExecutionController.class.getName());
 
   public Execution start(JobDescription job, Users user) throws IOException {
     Execution exec = null;
@@ -106,6 +119,36 @@ public class ExecutionController {
     return exec;
   }
 
+  public void kill(JobDescription job, Users user) throws IOException {
+
+    String appid = yarnApplicationstateFacade.findByAppname(job.getName())
+        .get(0)
+        .getApplicationid();
+
+    //Look for unique marker file which means it is a streaming job. Otherwise proceed with normal kill.
+    DistributedFileSystemOps udfso = null;
+    String username = hdfsUsersBean.getHdfsUserName(job.getProject(), user);
+    try {
+      udfso = dfs.getDfsOps(username);
+      String marker = Settings.getJobMarkerFile(job, appid);
+      if (udfso.exists(marker)) {
+        udfso.rm(new org.apache.hadoop.fs.Path(marker), false);
+
+      }
+    } catch (IOException ex) {
+      logger.log(Level.SEVERE, "Could not remove marker file for job:" + job.getName() + "with appId:" + appid, ex);
+    } finally {
+      if (udfso != null) {
+        udfso.close();
+      }
+    }
+
+    //WORKS FOR NOW BUT SHOULD EVENTUALLY GO THROUGH THE YARN CLIENT API
+    Runtime rt = Runtime.getRuntime();
+    Process pr = rt.exec(settings.getHadoopDir() + "/bin/yarn application -kill " + appid);
+
+  }
+  
   public void stop(JobDescription job, Users user, String appid) throws
           IOException {
     switch (job.getJobType()) {
